requirements.txt
discord.py==2.0.1
asyncio==3.4.3
apscheduler==3.9.1
requests==2.28.1
sqlalchemy==1.4.39
mysql-connector-python==8.0.29